fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@12308 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/xml/SUMOSAXAttributesImpl_Xerces.cpp
@@ -196,7 +196,7 @@ SUMOSAXAttributesImpl_Xerces::getName(int attr) const {
 
 void
 SUMOSAXAttributesImpl_Xerces::serialize(std::ostream& os) const {
-    for (int i = 0; i < myAttrs.getLength(); ++i) {
+    for (int i = 0; i < (int)myAttrs.getLength(); ++i) {
         os << " " << TplConvert<XMLCh>::_2str(myAttrs.getLocalName(i));
         os << "=\"" << TplConvert<XMLCh>::_2str(myAttrs.getValue(i)) << "\"";
     }
